+++ Source/Krypton Components/Krypton.Toolkit/Designers/Designers/KryptonLinkWrapLabelDesigner.cs
@@ -1,0 +1,46 @@
+﻿#region BSD License
+/*
+ * 
+ *  New BSD 3-Clause License (https://github.com/Krypton-Suite/Standard-Toolkit/blob/master/LICENSE)
+ *  Modifications by Peter Wagner(aka Wagnerp) & Simon Coghlan(aka Smurf-IV), et al. 2017 - 2023. All rights reserved. 
+ *  
+ */
+#endregion
+
+namespace Krypton.Toolkit
+{
+    internal class KryptonLinkWrapLabelDesigner : ControlDesigner
+    {
+        #region Identity
+
+        /// <summary>Initializes a new instance of the <see cref="KryptonLinkWrapLabelDesigner" /> class.</summary>
+        public KryptonLinkWrapLabelDesigner() =>
+            // The resizing handles around the control need to change depending on the
+            // value of the AutoSize and AutoSizeMode properties. When in AutoSize you
+            // do not get the resizing handles, otherwise you do.
+            AutoResizeHandles = true;
+
+        #endregion
+
+        #region Public Overrides
+        /// <summary>
+        ///  Gets the design-time action lists supported by the component associated with the designer.
+        /// </summary>
+        public override DesignerActionListCollection ActionLists
+        {
+            get
+            {
+                // Create a collection of action lists
+                DesignerActionListCollection actionLists = new()
+                {
+
+                    // Add the wrap label specific list
+                    new KryptonLinkWrapLabelActionList(this)
+                };
+
+                return actionLists;
+            }
+        }
+        #endregion
+    }
+}